Added GET, POST, PUT, Delete methods for CURD Operations:
GET "/?page=1" endpoint is created for fetching page details wherein each page is by default set to 10 that can be configurable in db.config.js file.
POST "/" endpoint is created for inserting a new record in DB, Data needs to be sent throught http request body in json format.
PUT "/:id" endpoint is created for updating any fields in the db, id needs to be given in url parameter and update key value should be sent in http request body.
DELETE "/:id" endpoint is created for deleteing a row from db, id needs to be sent in url parameter.